Change upload endpoint
Instead of using '/' as both frontpage and image upload endpoint,
disallow POST requests to root and return frontpage to GET requests.
As the user is redirected to frontpage after upload, the workflow
remains the same.

When using this approach, it should be more difficult for a user to
accidentally upload same image twice, as browser may suggest users to do
so in order to reload frontpage. This should not be an issue anymore.